+++ src/hash.rs
@@ -4,13 +4,13 @@
 //!
 //! We provide:
 //! - [`NoHash`]: does nothing -- only use on truly random keys.
-//! - [`IntHash`] = [`FxHash`], which does a single wrapping multiplication and should be good enough most of the time.
+//! - [`FastIntHash`] = [`FxHash`], which does a single wrapping multiplication and should be good enough most of the time.
 //! - [`StrongerIntHash`]: Use this when the input keys are very regular, eg `0..1000`. (But then why do you need an MPHF anyway?)
 //!                Does a `u128` multiplication, and xors the high and low word together, like xxh3. Then does one more multiplication. Not very scientific but 'it works'.
 //! - [`GxInt`]: GxHash, but with the type 'inlined' so that it optimized better.
 //! - [`Xxh3Int`]: Xxh3, but with the type 'inlined' so that it optimized better.
 //!
-//! In practice, prefer [`IntHash`] if it's good enough.
+//! In practice, prefer [`FastIntHash`] if it's good enough.
 //! Otherwise, fall back to [`StrongerIntHash`].
 //! If that still fails (which probably shouldn't happen) fall back to one of the two remaining
 //!
@@ -87,7 +87,7 @@ impl<Key: KeyT + ?Sized, H: core::hash::Hasher + Default + Clone + Sync> KeyHash
 // Aliases
 
 /// A slightly faster but weaker hash for sufficiently random integers. Uses [`fxhash::FxHasher64`].
-pub type IntHash = fxhash::FxHasher64;
+pub type FastIntHash = fxhash::FxHasher64;
 pub type FxHash = fxhash::FxHasher64;
 /// Type alias for xxhash (XXH3) hasher.
 ///